explode src/main.js to src/lib/* sub modules
