+++ src/lib/units.js
@@ -1,0 +1,13 @@
+'use strict'
+
+module.exports = {
+  getUnitValues: getUnitValues
+}
+
+function getUnitValues () {
+  return [
+    {text: 'pt', value: 'pt'},
+    {text: 'cm', value: 'cm'},
+    {text: 'mm', value: 'mm'}
+  ]
+}
